# springboot-microservices/user-auth-service/user-auth-service/target/classes/openapi.yaml
openapi: 3.0.1
info:
  description: "API for user registration, authentication, and user management"
  title: User Auth Service API
  version: 1.0.0
servers:
- url: http://localhost:8083
paths:
  /auth/register:
    post:
      requestBody:
        content:
          application/json:
            example:
              username: alice
              password: alice123
              email: alice@example.com
              role: ADMIN
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
        required: true
      responses:
        "200":
          description: User registered successfully
      summary: Register a new user
      x-content-type: application/json
      x-accepts:
      - application/json
  /auth/login:
    post:
      requestBody:
        content:
          application/json:
            example:
              username: alice
              password: alice123
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              example:
                token: eyJhbGciOiJIUzI1NiJ9...
              schema:
                $ref: '#/components/schemas/JwtResponse'
          description: JWT token response
      summary: Authenticate user and return JWT
      x-content-type: application/json
      x-accepts:
      - application/json
...
components:
  schemas:
    UserRegistrationRequest:
      example:
        password: password
        role: role
        email: email
        username: username
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
        role:
          type: string
      required:
      - email
      - password
      - role
      - username
      type: object
    UserLoginRequest:
      example:
        password: password
        username: username
      properties:
        username:
          type: string
        password:
          type: string
      required:
      - password
      - username
      type: object
    JwtResponse:
      example:
        token: token
      properties:
        token:
          type: string
      type: object
...